Fix Send/Receive Colors
This pull will fix the color issues on RPC as well as change the name of types from sent & Received to Send and Receive. 
There are two types of transactions: a send and a receive. There is no transaction type "Sent" and no transaction type "Received." Your wallet either performs a send or it performs a receive.